+++ src/sys/types/ino_t.c
@@ -1,6 +1,8 @@
 #include <sys/types.h>
 
-typedef unsigned long long int                                            ino_t;
+typedef                                                        unsigned long int
+	ino_t;
+
 /*
 POSIX(1)
 */
