+++ src/sys/types/off_t.c
@@ -1,6 +1,7 @@
 #include <sys/types.h>
 
-typedef long long int                                                     off_t;
+typedef                                                                 long int
+	off_t;
 /*
 POSIX(1)
 */
